\name{globpso-package}
\alias{globpso-package}
\alias{globpso}
\docType{package}
\title{
\packageTitle{globpso}
}
\description{
\packageDescription{globpso}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{globpso}
\packageIndices{globpso}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{globpso}

Maintainer: \packageMaintainer{globpso}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
